+++ packages/devon/src/commands/index.ts
@@ -2,5 +2,6 @@ export * as down from './down';
 export * as exec from './exec';
 export * as init from './init';
 export * as install from './install';
+export * as interact from './interact';
 export * as run from './run';
 export * as switch from './switch';