+++ packages/devon/src/commands/install/index.ts
@@ -1,12 +1,12 @@
 import { release, platform } from 'os';
 import { pathExists } from 'fs-extra';
 import { join, resolve } from 'path';
-import { green, red } from 'chalk';
 import sudo from 'sudo-prompt';
+import consola from 'consola';
 import ora from 'ora';
-import type { CommandHandler } from '@/types';
-import { config } from '@/config';
-import { cwd } from '@/helpers';
+import type { CommandHandler } from '../../types';
+import { config } from '../../config';
+import { cwd } from '../../helpers';
 
 export const run = async () => {
     const path = await cwd();
@@ -18,12 +18,15 @@ export const run = async () => {
     const certsExists = await pathExists(join(proxyPath, 'certs', 'rootCA.pem'));
 
     if (!certsExists) {
-        throw new Error('This monorepo has no certificates to install.');
+        throw new Error('this monorepo has no certificates to install');
     }
 
     // Check if they're running macOS 11+, since it has trouble installing via this method.
     if (platform() === 'darwin' && release().charAt(0) === '2') {
-        throw new Error(`macOS 11+ users must manually install the certificate chain - please run this command: \`sudo security add-trusted-cert -d -r trustRoot -k /Library/Keychains/System.keychain ${resolve(proxyPath, 'certs', 'rootCA.pem')}\``);
+        return [
+            `macOS 11+ users must manually install the certificate chain - please run this command: \`sudo security add-trusted-cert -d -r trustRoot -k /Library/Keychains/System.keychain ${resolve(proxyPath, 'certs', 'rootCA.pem')}\``,
+            'after this, you\'re all done!',
+        ];
     }
 
     return new Promise<void>((resolve, reject) => {
@@ -46,18 +49,24 @@ export const run = async () => {
 export const register: CommandHandler = ({ program }) => {
     program.command('install')
         .action(async () => {
-            const install = ora('Installing').start();
+            const install = ora('installing').start();
 
             try {
-                await run();
-
+                const result = await run();
                 install.stop();
-                console.log(green('Awesome! Everything is installed and ready-to-go.'));
-                console.log('');
-                console.log('Run `devon switch` to get started!');
+
+                if (typeof result === 'string' || Array.isArray(result)) {
+                    (Array.isArray(result) ? result : [result]).forEach(text => {
+                        consola.info(text);
+                    });
+                } else {
+                    consola.success('awesome! everything is installed and ready-to-go');
+                    consola.log('');
+                    consola.log('run `devon switch` to get started!');
+                }
             } catch (e) {
                 install.stop();
-                console.log(red((e as Error).message));
+                consola.error(e as Error);
             }
         });
 };