+++ packages/devon/src/commands/switch/index.ts
@@ -1,23 +1,35 @@
 import type { IDockerComposeOptions } from 'docker-compose';
 import compose from 'docker-compose';
-import { green, red } from 'chalk';
 import inquirer from 'inquirer';
+import consola from 'consola';
 import ora from 'ora';
-import { cwd, generateEnv, compileCompose, getServices, containerName } from '@/helpers';
-import { run as down } from '@/commands/down';
-import type { CommandHandler } from '@/types';
-import { config } from '@/config';
+import { cwd, generateEnv, compileCompose, getServices, containerName } from '../../helpers';
+import type { CommandHandler } from '../../types';
+import { run as down } from '../down';
+import { config } from '../../config';
 
-export const run = async (chosen: string[], env: string) => {
+export const run = async (chosen: string[], env: string, verbose = false) => {
+    const log = (msg: string) => {
+        if (verbose) {
+            consola.log(msg);
+        }
+    };
+
     // Fetch the definition and service configs.
     const { definition, services } = await getServices();
     const dir = await cwd();
 
-    console.log('');
-    const running = ora('Compiling Docker Compose').start();
+    consola.log('');
+    const running = ora('compiling docker-compose yaml').start();
+
+    if (verbose) {
+        running.stop();
+    }
 
     // Down existing compose if any.
+    log('downing previous containers (if any)');
     await down();
+    log('containers downed');
 
     // Compile the docker-compose config.
     const configs = [
@@ -40,6 +52,10 @@ export const run = async (chosen: string[], env: string) => {
 
     // Compile docker compose command config.
     const compiledConfig = await compileCompose(configs);
+
+    log('compiled docker-compose:');
+    log(compiledConfig);
+
     const composeConfig: IDockerComposeOptions = {
         configAsString: compiledConfig,
         cwd: dir,
@@ -51,12 +67,20 @@ export const run = async (chosen: string[], env: string) => {
     };
 
     // Trigger .env changes.
+    log('generating new .env files');
     await generateEnv(env);
+    log('.env files generated successfully');
 
     // Up the compose file.
     running.text = 'Starting containers';
+    log('starting containers');
+    const result = await compose.upAll(composeConfig);
 
-    await compose.upAll(composeConfig);
+    if (result.err) {
+        log(`failed to start containers: ${result.err}`);
+    } else {
+        log(`containers started successfully: ${result.out}`);
+    }
 
     config.set('currentCompose', {
         project: definition.project,
@@ -87,12 +111,13 @@ export const run = async (chosen: string[], env: string) => {
 
     // Done!
     running.stop();
-    console.log(green('Containers started successfully'));
+    consola.success('containers started');
 };
 
 export const register: CommandHandler = ({ program }) => {
     program.command('switch')
-        .action(async () => {
+        .option('-V, --verbose', 'output debug info and steps')
+        .action(async (opts: { verbose?: boolean }) => {
             try {
                 const { definition, services } = await getServices();
                 const existingServices = config.get('currentServices').find(item => item.project === definition.project);
@@ -104,7 +129,7 @@ export const register: CommandHandler = ({ program }) => {
                 }>([{
                     type: 'checkbox',
                     name: 'services',
-                    message: 'Which Services are you working on?',
+                    message: 'which Services are you working on?',
                     choices: Object.keys(services).filter(key => {
                         return services[key].definition.force !== true;
                     }).map(key => ({
@@ -115,17 +140,16 @@ export const register: CommandHandler = ({ program }) => {
                 }, {
                     type: 'list',
                     name: 'environment',
-                    message: 'Which environment should the other Services use?',
+                    message: 'which environment should the other services use?',
                     choices: definition.environments.map(env => ({
                         name: env.name,
                         value: env.name,
                     })),
                 }]);
 
-                await run(answers.services, answers.environment);
+                await run(answers.services, answers.environment, opts.verbose);
             } catch (e) {
-                // console.error(red((e as Error).message));
-                console.error(e);
+                consola.error(e as Error);
             }
         });
 };