+++ packages/devon/src/config/index.ts
@@ -1,6 +1,6 @@
 import Conf from 'conf';
+import type { ConfigSchema } from '../types';
 import pkg from '../../package.json';
-import type { ConfigSchema } from '@/types';
 
 export const config = new Conf<ConfigSchema>({
     projectName: 'devon',
@@ -14,29 +14,29 @@ export const config = new Conf<ConfigSchema>({
         dir: {
             type: 'string',
             default: '.devon',
-            description: 'Folder name for project-level storage',
+            description: 'folder name for project-level storage',
         },
         definitionFile: {
             type: 'string',
             default: '.devon.ts',
-            description: 'Filename for project-level config',
+            description: 'filename for project-level config',
         },
         serviceFile: {
             type: 'string',
             default: '.devconfig.ts',
-            description: 'Filename for service-level config',
+            description: 'filename for service-level config',
         },
         dataFolder: {
             type: 'string',
             default: '.data',
-            description: 'Folder name inside `dir` where Docker partitions can reside (Should be in .gitignore!)',
+            description: 'folder name inside `dir` where Docker partitions can reside (should be in .gitignore!)',
         },
         maxDepth: {
             type: 'number',
             default: 10,
             minimum: 1,
             maximum: 20,
-            description: 'When running a `devon` command, we can traverse up to this many levels to find the project definition',
+            description: 'when running a `devon` command, we can traverse up to this many levels to find the project definition',
         },
         currentCompose: {
             type: ['object', 'null'],
@@ -49,7 +49,7 @@ export const config = new Conf<ConfigSchema>({
                     type: 'string',
                 },
             },
-            description: 'The current compose config being served (Do not edit)',
+            description: 'the current compose config being served (do not edit!)',
         },
         currentServices: {
             type: 'array',
@@ -68,7 +68,7 @@ export const config = new Conf<ConfigSchema>({
                     },
                 },
             },
-            description: 'The current/last selected services mapped to each project for better UX (Do not edit)',
+            description: 'the current/last selected services mapped to each project for better UX (do not edit!)',
         },
     },
     migrations: {